requirements.txt
abbreviate==0.1.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
customtkinter==5.2.2
darkdetect==0.8.0
Deprecated==1.2.18
fastapi==0.116.2
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.35.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pefile==2023.2.7
pikepdf==9.11.0
pillow==11.3.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyenchant==3.3.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
PyMuPDF==1.26.4
PyPDF2==3.0.1
pypdfium2==5.0.0
pytesseract==0.3.13
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2025.9.1
reportlab==4.4.4
requests==2.32.5
sacremoses==0.1.1
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
